Make console_print_line be a scrollable

+++ kernel/console.h
@@ -1,3 +1,5 @@
+#include "types/types.h"
+
 void console_print_line(char* s);
-void console_set_cursor(int row, int col);
-void console_get_cursor(int *row, int *col);
+void console_set_cursor(uint16 row, uint16 col);
+void console_get_cursor(uint16 *row, uint16 *col);
